+++ NEWS.md
@@ -17,6 +17,7 @@
 * Using `dtw_lb` function now correctly warns about `pam.precompute` being `TRUE`.
 * Option to calculate pairwise distances with `SBD`, `LB_Keogh` and `LB_Improved`. See their respective notes. The distance function created for the `dtwclustFamily` slot also supports this.
 * Clusters are randomly re-initialized if they become empty at some iteration.
+* Now all included centroid functions recompute centers only if necessary.
 
 ## Version 1.3.0
 * Added the possibility to run several repetitions for partitional procedures, using different random starts each time by using the `doRNG` package